fuir: extract GeneratingFUIR.Clazz in separate file
